LayoutTests/platform/wpe/imported/w3c/web-platform-tests/encrypted-media/encrypted-media-default-feature-policy.https.sub-expected.txt

CONSOLE MESSAGE: line 9: ReferenceError: Can't find variable: run_all_fp_tests_allow_self

FAIL Untitled ReferenceError: Can't find variable: run_all_fp_tests_allow_self
